requirements.txt
tensorflow==2.15.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
flask==2.3.2
flask-cors==4.0.0
opencv-python==4.8.0.74
requests==2.31.0
locust==2.15.1
python-dotenv==1.0.0
joblib==1.3.1
scipy==1.11.2
werkzeug==2.3.7
